message and confirm moved to components folder
