+++ src/components/confirm/_confirm.scss
