+++ src/components/message/_message.scss
